backend/requirements.txt
fastapi~=0.115.2
uvicorn
pydantic~=2.9.2
pyjwt~=2.9.0
passlib~=1.7.4
bcrypt==4.0.1
sqlalchemy~=2.0.36
alembic~=1.13.3
requests~=2.32.3
pydantic-settings
python-dotenv~=1.0.1

